requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.8.0
django-filter==2.2.0
django-import-export==1.2.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.1.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
